Assigning HVA to Azure resources using tags (#17)
Adds support for reading certain tags on azure resource objects, making them into HVA_Tag objects, and adding them to the ad.json output file.

Resource tag should be formatted as 'C:1,I:2,A:3', with the numbers being the corresponding confidentiality, integrity and availability value for the given resource (with 0 being the lowest and 10 the highest).

Tag is case- and order-insensitive, and not all three need to be present (if one is absent, its value defaults to 0). So e.g. 'i:3,A:9', 'a:10', or 'a:0,i:9,c:5' are all valid tags.